feat: reuse prefix cache in compaction via structured messages + agent context
Compaction's LLM call previously used a completely different request
structure from normal agent turns:
- SystemPrompt: compact_system.md (~1KB) instead of agent prompt
- Messages: serialized text blob instead of structured multi-role messages
- No tools, no AgentContextPrefix (skills + AGENTS.md)

This meant zero prefix-cache hits — every compaction request started
from scratch at the first token.

This change aligns the compaction request structure with normal agent
turns so the entire prefix is served from cache:

  agent turn:  [sys_prompt] [tools] [ctx_prefix] [msg1...msgN]
  compaction:  [sys_prompt] [tools] [ctx_prefix] [msg1...msgK] [instruction]
                                                   ^^^^^^^^^^^^^^^ cached

Changes:
- pkg/context/conversion.go (new): shared ConvertMessagesToLLM,
  ConvertToolsToLLM, sanitizeToolCallProtocol, InsertBeforeFirstUserMessage
  so pkg/compact can reuse the exact same message serialization as the
  agent loop
- pkg/context/context.go: add AgentContextPrefix field to AgentContext
- pkg/agent/conversion.go: thin wrapper delegating to pkg/context
- pkg/compact/compact_summary.go: GenerateSummaryWithPrevious now accepts
  systemPrompt, contextPrefix, tools — builds a cache-aligned request
  with structured messages + agent prefix. Adds truncateForSummary guard
  (40% context window cap) to prevent oversized requests for large sessions
- pkg/compact/compact.go: pass ctx.SystemPrompt, ctx.AgentContextPrefix,
  ctx.Tools to summary generation
- cmd/ai/rpc_helpers.go: set ctx.AgentContextPrefix at startup
- Delete dead serialization code: serializeConversation,
  projectMessagesForSummary, truncateConversationToCharBudget,
  serializeSingleMessage, trimTextWithTail

Net: -640 lines, +308 lines.
